Refactored file structure slightly
