+++ packages/cw-orch-mock/src/core.rs
@@ -2,25 +2,21 @@ use std::{cell::RefCell, fmt::Debug, rc::Rc};
 
 use cosmwasm_std::{
     testing::{MockApi, MockStorage},
-    to_json_binary, Addr, Api, Binary, Coin, CosmosMsg, Empty, Event, Uint128, WasmMsg,
+    to_json_binary, Addr, Api, Binary, CosmosMsg, Empty, Event, WasmMsg,
 };
 use cw_multi_test::{
-    addons::{MockAddressGenerator, MockApiBech32},
-    ibc::IbcSimpleModule,
-    App, AppBuilder, AppResponse, BankKeeper, Contract, DistributionKeeper, Executor,
-    FailingModule, GovFailingModule, StakeKeeper, StargateFailing, WasmKeeper,
+    addons::MockApiBech32, ibc::IbcSimpleModule, App, AppResponse, BankKeeper, Contract,
+    DistributionKeeper, Executor, FailingModule, GovFailingModule, StakeKeeper, StargateFailing,
+    WasmKeeper,
 };
-use cw_utils::NativeBalance;
 use serde::Serialize;
 
 use super::state::MockState;
 use cw_orch_core::{
     contract::interface_traits::Uploadable,
-    environment::{BankQuerier, BankSetter, ChainState, IndexResponse, StateInterface, TxHandler},
+    environment::{ChainState, IndexResponse, StateInterface, TxHandler},
     CwEnvError,
 };
-
-use crate::queriers::bank::MockBankQuerier;
 
 pub type MockApp<A = MockApi> = App<
     BankKeeper,
@@ -90,243 +86,6 @@ impl<A: Api, S: StateInterface> Clone for MockBase<A, S> {
             sender: self.sender.clone(),
             state: self.state.clone(),
             app: self.app.clone(),
-        }
-    }
-}
-
-mod mock_api {
-    use super::*;
-
-    impl MockBase<MockApi, MockState> {
-        /// Create a mock environment with the default mock state.
-        pub fn new(sender: impl Into<String>) -> Self {
-            MockBase::<MockApi, MockState>::new_custom(sender, MockState::new())
-        }
-
-        pub fn new_with_chain_id(sender: impl Into<String>, chain_id: &str) -> Self {
-            let chain = MockBase::<MockApi, MockState>::new_custom(sender, MockState::new());
-            chain
-                .app
-                .borrow_mut()
-                .update_block(|b| b.chain_id = chain_id.to_string());
-
-            chain
-        }
-    }
-    impl<S: StateInterface> MockBase<MockApi, S> {
-        /// Create a mock environment with a custom mock state.
-        /// The state is customizable by implementing the `StateInterface` trait on a custom struct and providing it on the custom constructor.
-        pub fn new_custom(sender: impl Into<String>, custom_state: S) -> Self {
-            let state = Rc::new(RefCell::new(custom_state));
-            let app = Rc::new(RefCell::new(AppBuilder::new_custom().build(|_, _, _| {})));
-
-            Self {
-                sender: Addr::unchecked(sender),
-                state,
-                app,
-            }
-        }
-    }
-}
-
-mod bech32 {
-    use cw_orch_core::environment::DefaultQueriers;
-
-    use super::*;
-
-    impl MockBase<MockApiBech32, MockState> {
-        /// Create a mock environment with the default mock state.
-        pub fn new_bech32(prefix: &'static str) -> Self {
-            MockBech32::new_custom_bech32(prefix, MockState::new())
-        }
-
-        pub fn new_bech32_with_chain_id(prefix: &'static str, chain_id: &str) -> Self {
-            let chain = MockBech32::new_custom_bech32(prefix, MockState::new());
-            chain
-                .app
-                .borrow_mut()
-                .update_block(|b| b.chain_id = chain_id.to_string());
-
-            chain
-        }
-    }
-
-    impl<S: StateInterface> MockBase<MockApi, S> {
-        /// Set the bank balance of an address.
-        pub fn set_balance(
-            &self,
-            address: impl Into<String>,
-            amount: Vec<cosmwasm_std::Coin>,
-        ) -> Result<(), CwEnvError> {
-            self.app
-                .borrow_mut()
-                .init_modules(|router, _, storage| {
-                    router
-                        .bank
-                        .init_balance(storage, &Addr::unchecked(address.into()), amount)
-                })
-                .map_err(Into::into)
-        }
-
-        /// Adds the bank balance of an address.
-        pub fn add_balance(
-            &self,
-            address: impl Into<String>,
-            amount: Vec<cosmwasm_std::Coin>,
-        ) -> Result<(), CwEnvError> {
-            let addr = &Addr::unchecked(address.into());
-            let b = self.query_all_balances(addr.clone())?;
-            let new_amount = NativeBalance(b) + NativeBalance(amount);
-            self.app
-                .borrow_mut()
-                .init_modules(|router, _, storage| {
-                    router
-                        .bank
-                        .init_balance(storage, addr, new_amount.into_vec())
-                })
-                .map_err(Into::into)
-        }
-
-        /// Set the balance for multiple coins at once.
-        pub fn set_balances(
-            &self,
-            balances: &[(impl Into<String> + Clone, &[cosmwasm_std::Coin])],
-        ) -> Result<(), CwEnvError> {
-            self.app
-                .borrow_mut()
-                .init_modules(|router, _, storage| -> Result<(), CwEnvError> {
-                    for (addr, coins) in balances {
-                        router.bank.init_balance(
-                            storage,
-                            &Addr::unchecked(addr.clone()),
-                            coins.to_vec(),
-                        )?;
-                    }
-                    Ok(())
-                })
-        }
-
-        /// Query the (bank) balance of a native token for and address.
-        /// Returns the amount of the native token.
-        pub fn query_balance(
-            &self,
-            address: impl Into<String>,
-            denom: &str,
-        ) -> Result<Uint128, CwEnvError> {
-            Ok(self
-                .bank_querier()
-                .balance(address, Some(denom.to_string()))?
-                .first()
-                .map(|c| c.amount)
-                .unwrap_or_default())
-        }
-
-        /// Fetch all the balances of an address.
-        pub fn query_all_balances(
-            &self,
-            address: impl Into<String>,
-        ) -> Result<Vec<cosmwasm_std::Coin>, CwEnvError> {
-            self.bank_querier().balance(address, None)
-        }
-    }
-
-    impl<S: StateInterface> MockBase<MockApiBech32, S> {
-        pub fn create_account(&self, account_name: impl Into<String>) -> Addr {
-            self.app.borrow().api().addr_make(&account_name.into())
-        }
-    }
-
-    impl Default for MockBase<MockApiBech32, MockState> {
-        fn default() -> Self {
-            MockBase::<MockApiBech32, MockState>::new_custom_bech32("mock", MockState::new())
-        }
-    }
-
-    impl<S: StateInterface> MockBase<MockApiBech32, S> {
-        /// Create a mock environment with a custom mock state.
-        /// The state is customizable by implementing the `StateInterface` trait on a custom struct and providing it on the custom constructor.
-        pub fn new_custom_bech32(prefix: &'static str, custom_state: S) -> Self {
-            let state = Rc::new(RefCell::new(custom_state));
-            let app = Rc::new(RefCell::new(
-                AppBuilder::new_custom()
-                    .with_api(MockApiBech32::new(prefix))
-                    .with_wasm(WasmKeeper::default().with_address_generator(MockAddressGenerator))
-                    .build(|_, _, _| {}),
-            ));
-
-            // We create an address internally
-            let sender = app.borrow().api().addr_make("sender");
-
-            Self { sender, state, app }
-        }
-    }
-
-    impl<S: StateInterface> MockBech32<S> {
-        /// Set the bank balance of an address.
-        pub fn set_balance(
-            &self,
-            address: &Addr,
-            amount: Vec<cosmwasm_std::Coin>,
-        ) -> Result<(), CwEnvError> {
-            self.app
-                .borrow_mut()
-                .init_modules(|router, _, storage| {
-                    router.bank.init_balance(storage, address, amount)
-                })
-                .map_err(Into::into)
-        }
-
-        /// Adds the bank balance of an address.
-        pub fn add_balance(
-            &self,
-            address: &Addr,
-            amount: Vec<cosmwasm_std::Coin>,
-        ) -> Result<(), CwEnvError> {
-            let addr = &address;
-            let b = self.query_all_balances(addr)?;
-            let new_amount = NativeBalance(b) + NativeBalance(amount);
-            self.app
-                .borrow_mut()
-                .init_modules(|router, _, storage| {
-                    router
-                        .bank
-                        .init_balance(storage, addr, new_amount.into_vec())
-                })
-                .map_err(Into::into)
-        }
-
-        /// Set the balance for multiple coins at once.
-        pub fn set_balances(
-            &self,
-            balances: &[(&Addr, &[cosmwasm_std::Coin])],
-        ) -> Result<(), CwEnvError> {
-            self.app
-                .borrow_mut()
-                .init_modules(|router, _, storage| -> Result<(), CwEnvError> {
-                    for (addr, coins) in balances {
-                        router.bank.init_balance(storage, addr, coins.to_vec())?;
-                    }
-                    Ok(())
-                })
-        }
-
-        /// Query the (bank) balance of a native token for and address.
-        /// Returns the amount of the native token.
-        pub fn query_balance(&self, address: &Addr, denom: &str) -> Result<Uint128, CwEnvError> {
-            Ok(self
-                .bank_querier()
-                .balance(address, Some(denom.to_string()))?
-                .first()
-                .map(|c| c.amount)
-                .unwrap_or_default())
-        }
-
-        /// Fetch all the balances of an address.
-        pub fn query_all_balances(
-            &self,
-            address: &Addr,
-        ) -> Result<Vec<cosmwasm_std::Coin>, CwEnvError> {
-            self.bank_querier().balance(address, None)
         }
     }
 }
@@ -488,29 +247,6 @@ impl<A: Api, S: StateInterface> TxHandler for MockBase<A, S> {
     }
 }
 
-impl<S: StateInterface> BankSetter for Mock<S> {
-    type T = MockBankQuerier<MockApi>;
-
-    fn set_balance(
-        &mut self,
-        address: impl Into<String>,
-        amount: Vec<Coin>,
-    ) -> Result<(), <Self as TxHandler>::Error> {
-        (*self).set_balance(address, amount)
-    }
-}
-impl<S: StateInterface> BankSetter for MockBech32<S> {
-    type T = MockBankQuerier<MockApiBech32>;
-
-    fn set_balance(
-        &mut self,
-        address: impl Into<String>,
-        amount: Vec<Coin>,
-    ) -> Result<(), <Self as TxHandler>::Error> {
-        (*self).set_balance(&Addr::unchecked(address), amount)
-    }
-}
-
 #[cfg(test)]
 mod test {
 
@@ -519,7 +255,7 @@ mod test {
         StdResult, Uint128,
     };
     use cw_multi_test::ContractWrapper;
-    use cw_orch_core::environment::{DefaultQueriers, QueryHandler};
+    use cw_orch_core::environment::{BankQuerier, DefaultQueriers, QueryHandler};
     use serde::Serialize;
     use speculoos::prelude::*;
